% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smmit.R
\name{smmit}
\alias{smmit}
\title{Smmit: single-cell multi-sample and multi-omics integration}
\usage{
smmit(obj, mode = "RNA_ATAC")
}
\arguments{
\item{obj}{A list of named Seurat objects. Each Seurat object contains single-cell multi-omics data from a single sample, processed using the standard Seurat or Signac pipeline. The name of each element is the sample name.}

\item{mode}{Either RNA_ATAC or RNA_ADT. RNA_ATAC performs integration for single-cell multi-omics data that jointly profile gene expression and chromatin accessibility. RNA_ADT performs integration for single-cell multi-omics data that jointly profile gene expression and protein abundances.}
}
\value{
A Seurat object. The UMAP that integrates both across samples and modalities is stored in the wsnnumap reduction. The Seurat object can be directly used in downstream analysis such as cell clustering, cell type identification, and differential analysis.
}
\description{
Smmit performs integration both across samples and modalities to produce a single UMAP space. It first uses harmony to integrate across samples and then uses Seurat weighted nearest neighbor function to integrate across modalities.
}
\author{
Changxin Wan, Zhicheng Ji<zhicheng.ji@duke.edu>
}
